+++ LojaHabitavel.java
@@ -10,9 +10,9 @@ public class LojaHabitavel extends Loja
         apartamento = new Apartamento();
     }
     
-    public LojaHabitavel(String id, String rua, double precoPedido, double precoMinimo, int area,
-                         boolean temWC, String tipoNegocio, int numDaPorta, Apartamento apartamento){
-        
+    public LojaHabitavel(String id, String rua, int precoPedido, int precoMinimo, int area,
+                         boolean temWC, String tipoNegocio, int numDaPorta, Apartamento apartamento)
+    {
         super(id, rua, precoPedido, precoMinimo, area, temWC, tipoNegocio, numDaPorta);
         setApartamento(apartamento);
     }
